+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Changed
 - Naturalness vector artifact is renamed to "Path Greenness" ([#283](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/283))
+- Raise a `ClimatoologyUserError` if there's an error with Ohsome ([#287](https://gitlab.heigit.org/climate-action/plugins/walkability/-/issues/287))
 
 ### Removed
 - all the shenanigans around accessing a private climatoology repository because that is now public